face-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==1.26.4
pillow==10.4.0
insightface==0.7.3
onnxruntime==1.18.1
